26/Oct/2022 - Code formatting

+++ controllers/genres.js
@@ -8,19 +8,19 @@ module.exports = (dependencies) => {
 
         const genres = await dependencies.models.genre.find();
 
-        const genres_getall= [];
-        
+        const genres_getall = [];
+
         try {
-        for(const genre of genres) {
-            genres_getall.push({
-                id: genre.id,
-                name: genre.name
-            });
+            for (const genre of genres) {
+                genres_getall.push({
+                    id: genre.id,
+                    name: genre.name
+                });
+            }
+            response.status(200).send(genres_getall);
+        } catch (err) {
+            console.error(err);
         }
-        response.status(200).send(genres_getall);
-    }catch(err) {
-        console.error(err);
-    }
     });
 
     router.get('/:genre_id', async (request, response, next) => {
@@ -35,17 +35,17 @@ module.exports = (dependencies) => {
             #swagger.tags = ['Genre']
             #swagger.description = 'Get a specific genre by genre_id'
         */
-            const genre = await dependencies.models.genre.findOne({_id: request.params.genre_id});
+        const genre = await dependencies.models.genre.findOne({ _id: request.params.genre_id });
 
-            if(!genre) {
-                return response.status(404).send('Genre not found');
-            }
-            const genres_processed = {};
+        if (!genre) {
+            return response.status(404).send('Genre not found');
+        }
+        const genres_processed = {};
 
-            genres_processed.id = genre._id;
-            genres_processed.name = genre.name;
+        genres_processed.id = genre._id;
+        genres_processed.name = genre.name;
 
-            response.status(200).send(genres_processed);
+        response.status(200).send(genres_processed);
     });
 
     // The second callback over here makes authentication required for this endpoint
@@ -67,21 +67,21 @@ module.exports = (dependencies) => {
                 }
             }
         */
-        
-            if (!request.body.name) {
-                return response.status(400).send('One or more of the required fields are missing.');
-            } else {
-                //not sure if this is needed so I am commenting it out
-                //next();
-            }
-            const new_genre = await dependencies.models.genre.create(request.body);
 
-            if (!new_genre) {
-                response.status(500).send('An error occured.');
-            } else {
-                response.status(201).send('Successfully added genre. ID: ' + new_genre._id);
-            }
-        });
+        if (!request.body.name) {
+            return response.status(400).send('One or more of the required fields are missing.');
+        } else {
+            //not sure if this is needed so I am commenting it out
+            //next();
+        }
+        const new_genre = await dependencies.models.genre.create(request.body);
+
+        if (!new_genre) {
+            response.status(500).send('An error occured.');
+        } else {
+            response.status(201).send('Successfully added genre. ID: ' + new_genre._id);
+        }
+    });
 
     router.put('/:genre_id', dependencies.requires_authentication(), async (request, response, next) => {
         /*
@@ -107,9 +107,12 @@ module.exports = (dependencies) => {
                 }
             }
         */
-        const result = await dependencies.models.genre.updateOne({_id: request.params.genre_id}, request.body);
+        if (!request.params.genre_id) {
+            return response.status(400).send('The genre_id is a required field.');
+        }
+        const result = await dependencies.models.genre.updateOne({ _id: request.params.genre_id }, request.body);
         response.status(200).send('Updated genre with ID: ' + request.params.genre_id);
-    }); 
+    });
 
     router.delete('/:genre_id', dependencies.requires_authentication(), async (request, response, next) => {
         /*
@@ -124,22 +127,26 @@ module.exports = (dependencies) => {
             #swagger.description = 'Deletes a genre specified by genre_id'
         */
 
-            const result = await dependencies.models.genre.deleteOne({_id: request.params.genre_id});
+        if (!request.params.genre_id) {
+            return response.status(400).send('The genre_id is a required field.');
+        }
 
-            return response.status(200).send('Deleted genre with ID: ' + request.params.genre_id);
-            
-            /*
-            const genre = dependencies.models.genre.findOne({_id: request.params.genre_id});
-            if(!genre) {
-                return response.status(404).send('Genre not found');
-            }
-            const genres_processed = {};
+        const result = await dependencies.models.genre.deleteOne({ _id: request.params.genre_id });
 
-            genres_processed.id = genre._id;
-            genres_processed.name = genre.name;
+        return response.status(200).send('Deleted genre with ID: ' + request.params.genre_id);
 
-            response.status(200).send(genres_processed);
-            */
+        /*
+        const genre = dependencies.models.genre.findOne({_id: request.params.genre_id});
+        if(!genre) {
+            return response.status(404).send('Genre not found');
+        }
+        const genres_processed = {};
+
+        genres_processed.id = genre._id;
+        genres_processed.name = genre.name;
+
+        response.status(200).send(genres_processed);
+        */
     });
 
     return router;
